BJS2-78003. Unit-тесты для задачи с подписками - Тесты для service, filter, validator - Правки по PR

+++ src/main/java/school/faang/user_service/controller/subscription/SubscriptionController.java
@@ -82,7 +82,7 @@ public class SubscriptionController {
     }
 
     @GetMapping("{userId}/subscriptions/followingCount")
-    public ResponseEntity<Integer> getFollowingCount(@PathVariable("userId") long followerId) {
+    public ResponseEntity<Long> getFollowingCount(@PathVariable("userId") long followerId) {
         return new ResponseEntity<>(
                 subscriptionService.getFollowingCount(followerId),
                 HttpStatus.OK
